refactor: move spider, surround to label.position
