+++ src/label-transform/index.ts
@@ -1,13 +1,9 @@
 export { OverlapHide } from './overlapHide';
 export { OverlapDodgeY } from './overlapDodgeY';
-export { Spider } from './spider';
-export { Surround } from './surround';
 export { ContrastReverse } from './contrastReverse';
 export { OverflowHide } from './overflowHide';
 
 export type { OverlapHideOptions } from './overlapHide';
 export type { OverlapDodgeYOptions } from './overlapDodgeY';
-export type { SpiderOptions } from './spider';
-export type { SurroundOptions } from './surround';
 export type { ContrastReverseOptions } from './contrastReverse';
 export type { OverflowHideOptions } from './overflowHide';
