+++ src/shape/label/position/bottom.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as bottom } from './default';
