+++ src/shape/label/position/bottomLeft.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as bottomLeft } from './default';
