+++ src/shape/label/position/bottomRight.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as bottomRight } from './default';
