+++ src/shape/label/position/inside.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as inside } from './default';
