+++ src/shape/label/position/left.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as left } from './default';
