+++ src/shape/label/position/outside.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as outside } from './default';
