+++ src/shape/label/position/right.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as right } from './default';
