+++ src/shape/label/position/top.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as top } from './default';
