+++ src/shape/label/position/topLeft.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as topLeft } from './default';
