+++ src/shape/label/position/topRight.ts
@@ -1,0 +1,1 @@
+export { getDefaultStyle as topRight } from './default';
